refactor(AppHost) Refactor Kafka topic management in extensions
- Moved event publishing and consuming topic constants to `ResourceExtensions.cs`.
- Moved `SetupKafka` method to `ResourceExtensions.cs` to configure Kafka for microservices.
- Moved `GetTopicName` method to `ResourceExtensions.cs` for improved organization.
- Overall, these changes enhance the structure and maintainability of Kafka topic management code.
