+++ Aloha/Aloha.AppHost/Extensions/ExternalServiceRegistrationExtentions.cs
@@ -1,17 +1,10 @@
 using Microsoft.Extensions.Configuration;
+using static Aloha.AppHost.Extensions.ResourceExtensions;
 
 namespace Aloha.AppHost.Extensions;
 
 public static class ApplicationServiceExtensions
 {
-    #region Constants
-    private static class Consts
-    {
-        public const string Env_EventPublishingTopics = "EVENT_PUBLISHING_TOPICS";
-        public const string Env_EventConsumingTopics = "EVENT_CONSUMING_TOPICS";
-    }
-    #endregion
-
     #region External Service Registration
     public static IDistributedApplicationBuilder AddApplicationServices(this IDistributedApplicationBuilder builder)
     {
@@ -77,6 +70,7 @@ public static class ApplicationServiceExtensions
         return builder;
     }
 
+    #region CreateKafkaTopics Implementation
     private static async Task CreateKafkaTopics(ResourceReadyEvent @event, KafkaServerResource kafkaResource, CancellationToken ct)
     {
         var logger = @event.Services.GetRequiredService<ILogger<Program>>();
@@ -102,28 +96,6 @@ public static class ApplicationServiceExtensions
             logger.LogError(ex, "An error occurred creating topics");
         }
     }
+    #endregion
 
-    private static string GetTopicName<TProject>(string postfix = "") => $"{typeof(TProject).Name.Replace('_', '-')}{(string.IsNullOrEmpty(postfix) ? "" : $"-{postfix}")}";
-
-    /// <summary>
-    /// Cau hinh Kafka cho cac microservice - them tham chieu, dependencies, dang ky cac Publisher va Consumer cho cac topic.
-    /// </summary>
-    /// <typeparam name="TProject">Xac dinh service su dung Extension Method nay</typeparam>
-    /// <param name="serviceBuilder">Builder cua service can cau hinh</param>
-    /// <param name="kafkaResource">Builder cua Kafka Server Resource</param>
-    /// <param name="consumingFromServices">Danh sach cac service se tieu thu topic</param>
-    /// <returns>Builder da duoc cau hinh voi cac tham so can thiet</returns>
-    private static IResourceBuilder<ProjectResource> SetupKafka<TProject>(
-        this IResourceBuilder<ProjectResource> serviceBuilder,
-        IResourceBuilder<KafkaServerResource> kafkaResource,
-        params string[] consumingFromServices)
-    {
-        var topicNames = consumingFromServices.Select(s => s.Replace("_", "-")).ToArray();
-
-        return serviceBuilder
-        .WithEnvironment(Consts.Env_EventPublishingTopics, GetTopicName<TProject>())
-        .WithEnvironment(Consts.Env_EventConsumingTopics, string.Join(',', topicNames))
-        .WithReference(kafkaResource)
-        .WaitFor(kafkaResource);
-    }
 }
